docs: update decompilation progress metrics and function matching
Increased the decompiled C functions to **54.0%** (342/633) and bytes to **25.4%** (22,890/90,272). Updated various documentation files to reflect the latest progress, including improvements in function matching and semantic C scores. Notable updates include refined notes for several functions, the addition of new patterns, and corrections to existing entries, enhancing clarity and accuracy in the decompilation process.

+++ src/matched/sub_08072F94.c
@@ -1,7 +1,7 @@
 #include "global.h"
 
 // @ 0x08072f94
-u8 sub_08072F94(void)
+u32 sub_08072F94(void)
 {
     u32 tmp[2];
     u32 *ptr;
